chore: refactor into crypto module
